# be-devtrack/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DevTrack API",
        "description": "API untuk Project and Issue Tracking Tool",
        "contact": {
            "email": "wasis1219@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        },
        {
            "url": "https://api.devtrack.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/projects/{id}/report": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get project report",
                "description": "Get report with all issues and work logs for a project",
                "operationId": "getProjectReport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project report",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "project": {
                                            "type": "object"
                                        },
                                        "issues": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "type": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "priority": {
                                                        "type": "integer"
                                                    },
                                                    "assigned_engineer": {
                                                        "type": "object"
                                                    },
                                                    "total_working_hours": {
                                                        "type": "number"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total_hours": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/issues/{issueId}/work-logs": {
            "get": {
                "tags": [
                    "Work Logs"
                ],
                "summary": "Get work logs for issue",
                "description": "Get work logs for an issue",
                "operationId": "getWorkLogs",
                "parameters": [
                    {
                        "name": "issueId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Work logs list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Work Logs"
                ],
                "summary": "Add work log",
                "description": "Assigned engineer can log work on the issue",
                "operationId": "createWorkLog",
                "parameters": [
                    {
                        "name": "issueId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hours",
                                    "logged_at"
                                ],
                                "properties": {
                                    "hours": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 4.5
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "logged_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Work log created"
                    },
                    "403": {
                        "description": "Only assigned engineer can log work"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/work-logs/{id}": {
            "get": {
                "tags": [
                    "Work Logs"
                ],
                "summary": "Get work log details",
                "description": "Get work log details",
                "operationId": "getWorkLogById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Work log details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Work Logs"
                ],
                "summary": "Update work log",
                "description": "User who logged work can update it",
                "operationId": "updateWorkLog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "hours": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "logged_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Work log updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Work Logs"
                ],
                "summary": "Delete work log",
                "description": "Delete work log",
                "operationId": "deleteWorkLog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Work log deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}